moves multiSliding into NGram
